count down,  unable to set alarm error, retry scheme.
The startup delay is now a countdown instead of a countup.

Status from the call to set an alarm is now checked.  There is an associated progress constant.

The retry scheme is more robust.  Previously, if an NTP UDP packet was lost, the system failed to retry.
Now if udp isn't received within 30 seconds, a retry of the whole process is restarted.

Refactoring to make NTP_T member names more meaningful.

There are more debug printfs now during the testing phase