Wrote tests to induce packet loss.  Revised recovery strategy.
The test for loss of an NTP response showed up a problem. When the alarm timed out, code would hang.  I couldn't understand why it would hang.  But I removed the alarm, since there was already a retry mechanism.   When ntp_ask_for_time() is called, it sets the time at which to retry.  If an ntp packet is received, the retry is disabled.